+++ gallery-backend/src/public/db/tree_snapshot/new.rs
@@ -8,7 +8,7 @@ use super::TreeSnapshot;
 static TREE_SNAPSHOT_IN_DISK: LazyLock<redb::Database> =
     LazyLock::new(|| redb::Database::create("./db/temp_db.redb").unwrap());
 
-static TREE_SNAPSHOT_IN_MEMORY: LazyLock<DashMap<u128, Vec<ReducedData>>> =
+static TREE_SNAPSHOT_IN_MEMORY: LazyLock<DashMap<i64, Vec<ReducedData>>> =
     LazyLock::new(|| DashMap::new());
 
 impl TreeSnapshot {
